+++ Character/Class/General/OffhandAttack.cpp
@@ -44,34 +44,34 @@ void OffhandAttack::calculate_damage() {
         return;
     }
 
-    float damage_dealt = pchar->get_random_non_normalized_oh_dmg() * 0.5;
+    double damage_dealt = pchar->get_random_non_normalized_oh_dmg() * 0.5;
 
     if (result == AttackResult::CRITICAL) {
         damage_dealt *= 2;
         pchar->melee_oh_white_critical_effect();
-        add_crit_dmg(round(damage_dealt));
+        add_crit_dmg(static_cast<int>(round(damage_dealt)));
         return;
     }
     if (result == AttackResult::GLANCING) {
         damage_dealt *= roll->get_glancing_blow_dmg_penalty(oh_wpn_skill);
         pchar->melee_oh_white_hit_effect();
-        add_glancing_dmg(round(damage_dealt));
+        add_glancing_dmg(static_cast<int>(round(damage_dealt)));
         return;
     }
 
     pchar->melee_oh_white_hit_effect();
-    add_hit_dmg(round(damage_dealt));
+    add_hit_dmg(static_cast<int>(round(damage_dealt)));
 }
 
-float OffhandAttack::get_next_expected_use() const {
+double OffhandAttack::get_next_expected_use() const {
     return next_expected_use;
 }
 
-void OffhandAttack::update_next_expected_use(const float haste_change) {
+void OffhandAttack::update_next_expected_use(const double haste_change) {
     assert(haste_change > 0.001 || haste_change < -0.001);
 
-    float curr_time = pchar->get_engine()->get_current_priority();
-    float remainder_after_haste_change = (next_expected_use - curr_time);
+    double curr_time = pchar->get_engine()->get_current_priority();
+    double remainder_after_haste_change = (next_expected_use - curr_time);
 
     if (remainder_after_haste_change < -0.0001)
         return;
