+++ Character/Class/General/Procs/ExtraAttackInstantProc.cpp
@@ -8,7 +8,7 @@
 ExtraAttackInstantProc::ExtraAttackInstantProc(Engine* engine, Character* pchar, CombatRoll* roll,
                                                QString proc_name,
                                                QVector<ProcInfo::Source> proc_sources,
-                                               const float proc_rate, const int num_attacks) :
+                                               const double proc_rate, const int num_attacks) :
     Proc(proc_name, proc_rate, 0, false, QVector<Proc*>(), proc_sources, engine, pchar, roll),
     extra_attack_buff(nullptr),
     num_attacks(num_attacks)
