+++ Character/Class/Shaman/Shaman.cpp
@@ -67,7 +67,7 @@ int Shaman::get_ap_per_agi() const {
     return 1;
 }
 
-float Shaman::global_cooldown() const {
+double Shaman::global_cooldown() const {
     return 1.5;
 }
