+++ Character/Class/Warlock/Warlock.h
@@ -21,7 +21,7 @@ public:
     int get_spirit_modifier() const override;
     int get_agi_needed_for_one_percent_phys_crit() const override;
     int get_int_needed_for_one_percent_spell_crit() const override;
-    float global_cooldown() const override;
+    double global_cooldown() const override;
 
     int get_ap_per_strength() const override;
     int get_ap_per_agi() const override;
