+++ Character/Race/Races/Human.cpp
@@ -29,11 +29,11 @@ int Human::get_base_spirit(void) const {
     return 20;
 }
 
-float Human::get_int_multiplier() const {
+double Human::get_int_multiplier() const {
     return 1.0;
 }
 
-float Human::get_spirit_multiplier() const {
+double Human::get_spirit_multiplier() const {
     return 1.05;
 }
